Refactor Product Validation and DTO Implementation
- Implemented string validation for required fields in Product and ProductDTO.
- Added JavaDocs to improve code clarity and documentation.
- Refined the initial structure of the code for better maintainability.
- Enhanced validation logic to ensure proper field constraints in Product and ProductDTO.

+++ src/test/java/com/stores/Inventory/Unittest/ProductServiceTests.java
@@ -37,7 +37,7 @@ public class ProductServiceTests {
         MockitoAnnotations.openMocks(this);
         
         // Create a Product instance to be returned by productDTO.toProduct()
-        product = new Product("car", "BMW", (float) 10000.0, 2);
+        product = new Product("car", "BMW", (Double) 10000.0, 2);
         productList = new ArrayList<Product>();
         productList.add(product);
